feat(p1): add developer guardrails with ruff, make check, and CI integration
- Add ruff as test dependency; configure lint rules in pyproject.toml
- Add make lint, make lint-fix, make fix, make format, make check, make package-smoke
- Update make test and make install-local to use uv run --python $(UV_PYTHON)
- Add guardrails CI job running make UV_PYTHON=3.13 check
- Add requests and defusedxml as explicit runtime dependencies
- Bump version to 0.10.2 and add CHANGELOG entry
- Add uv.lock; update .gitignore with .venv/ and dist/
- Move broad CLI clipboard tests to tests/integration/ytt/
- Update README with local verification workflow
- Mark all P1 tasks as complete in docs/plans/2026-06-22-p1-developer-guardrails.md
- Update AGENTS.md to require running local gates before finishing iteration
