+++ .github/workflows/ci.yml
@@ -7,6 +7,25 @@ on:
       - main
 
 jobs:
+  guardrails:
+    name: lint, tests, package smoke
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Check out repository
+        uses: actions/checkout@v4
+
+      - name: Set up Python 3.13
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.13"
+
+      - name: Set up uv
+        uses: astral-sh/setup-uv@v6
+
+      - name: Run local verification suite
+        run: make UV_PYTHON=3.13 check
+
   tox:
     name: tox (${{ matrix.python-version }})
     runs-on: ubuntu-latest
