+++ AGENTS.md
@@ -6,9 +6,16 @@
 - YouTube URL parsing lives in `src/ytt/domain/value_objects.py`; unit tests belong in `tests/unit/ytt/domain/test_value_objects.py`.
 - For unit tests, mirror the `src/` directory inside `tests/unit/` and name files `test_<source-file>.py` (add a suffix only when splitting large suites).
 
-## Process: PRD → Spec → Plan → Deliver
+## Process Options: Full Flow or Standalone Plan
 
-Follow this lightweight flow for any non-trivial feature or change. Keep documents short but crisp—optimize for quick iteration by an agent.
+Use one of two tracks before implementation:
+
+- Full flow: use PRD → Spec → Plan → Deliver for non-trivial features or changes where the extra design context improves implementation.
+- Standalone plan: use a dated plan only for roadmap, priority, audit, planning-only, or one-off implementation work where creating matching PRD and spec files would add overhead without improving clarity.
+
+Keep documents short but crisp—optimize for quick iteration by an agent. Do not add PRD/Spec placeholders to standalone dated plans.
+
+### Full Flow: PRD → Spec → Plan → Deliver
 
 1) PRD (what & why)
 - Location: `docs/prds/`
@@ -27,7 +34,7 @@ Follow this lightweight flow for any non-trivial feature or change. Keep documen
 - Contents: task breakdown (checklist), sequencing/milestones, dependencies, owner(s), risks/mitigations, Definition of Done.
 - Must link to both PRD and Spec; keep reciprocal links updated (PRD ↔ Plan, Spec ↔ Plan).
 
-Standalone plans that are not part of a synchronized PRD/Spec/Plan set should also live in `docs/plans/`, but must use a date prefix instead of the numbered `NNN-` prefix: `YYYY-MM-DD-<slug>.md`. Use this for roadmap, priority, audit, or planning-only documents where creating matching PRD and spec files would add overhead without improving implementation clarity. When creating dated plan filenames, run `date -I` and use that value.
+Standalone plans that are not part of a synchronized PRD/Spec/Plan set should also live in `docs/plans/`, but must use a date prefix instead of the numbered `NNN-` prefix: `YYYY-MM-DD-<slug>.md`. When creating dated plan filenames, run `date -I` and use that value.
 
 4) Deliver (implement)
 - Create branch: `feat/NNN-<slug>` and implement per plan.
@@ -35,6 +42,7 @@ Standalone plans that are not part of a synchronized PRD/Spec/Plan set should al
 - If behavior/packaging/public interface changes: bump version in `pyproject.toml` and add a `CHANGELOG.md` entry.
 - Add ADR(s) in `docs/adrs/` when introducing notable structural/architectural decisions.
 - For any `last-updated` style metadata, use `date -I`.
+- Before finishing an implementation iteration, run the relevant local gates; default to `make fix` followed by `make check` unless the change clearly needs a narrower or broader command.
 - Open a PR referencing the PRD/Spec/Plan; ensure CI/tests pass.
 
 Quick checklist
@@ -44,6 +52,7 @@ Quick checklist
 - [ ] Feature branch created (`feat/NNN-<slug>`)
 - [ ] Implementation matches plan; docs kept in sync
 - [ ] Tests added/updated
+- [ ] Local gates run (`make fix` then `make check` by default)
 - [ ] Version + changelog updated if required
 - [ ] ADR(s) added if needed
 - [ ] PR opened with links to PRD/Spec/Plan
