+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "ytt"
-version = "0.10.1"
+version = "0.10.2"
 description = "A simple CLI tool to fetch YouTube video transcripts."
 readme = "README.md"
 requires-python = ">=3.12"
@@ -20,11 +20,16 @@ dependencies = [
     "youtube-transcript-api~=1.2",
     "appdirs~=1.4.4",
     "pyperclip~=1.9.0",
+    "requests~=2.32",
+    "defusedxml~=0.7.1",
 ]
 
 [project.optional-dependencies]
 test = [
+    "build>=1.2",
     "pytest>=8.0",
+    "ruff>=0.8",
+    "setuptools>=61.0",
 ]
 
 [project.scripts]
@@ -41,3 +46,14 @@ include = ["ytt*"]
 [tool.pytest.ini_options]
 pythonpath = ["src"]
 testpaths = ["tests"]
+
+[tool.ruff]
+target-version = "py312"
+line-length = 100
+extend-exclude = [
+    "build",
+    "*.egg-info",
+]
+
+[tool.ruff.lint]
+select = ["E4", "E7", "E9", "F"]
