+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 requires = tox>=4
-env_list = py312, py313
+env_list = lint, py312, py313, pkg
 isolated_build = true
 
 [testenv]
@@ -10,3 +10,19 @@ commands =
     ytt --help
     ytt --version
     python -c "import os, pathlib, tempfile; from importlib import metadata; os.chdir(tempfile.mkdtemp()); import ytt, ytt.domain; path = pathlib.Path(ytt.__file__).resolve(); print(path); assert 'site-packages' in str(path); assert metadata.version('ytt') == ytt.__version__"
+
+[testenv:lint]
+extras = test
+commands =
+    ruff check .
+
+[testenv:format]
+extras = test
+commands =
+    ruff format --check .
+
+[testenv:pkg]
+extras = test
+commands =
+    python -c "import shutil; [shutil.rmtree(path, ignore_errors=True) for path in ('build', 'dist', 'src/ytt.egg-info', 'ytt.egg-info', 'youtube_transcript_tool.egg-info')]"
+    python -m build --sdist --wheel --no-isolation
